created deploy/retrieve in util
