+++ packages/salesforcedx-vscode-core/package.json
@@ -27,7 +27,7 @@
   "dependencies": {
     "@salesforce/core": "2.4.1",
     "@salesforce/salesforcedx-utils-vscode": "48.19.0",
-    "@salesforce/source-deploy-retrieve": "1.0.10",
+    "@salesforce/source-deploy-retrieve": "1.0.11",
     "adm-zip": "0.4.13",
     "applicationinsights": "1.0.6",
     "glob": "^7.1.2",
@@ -88,8 +88,7 @@
   "main": "./out/src",
   "contributes": {
     "viewsContainers": {
-      "activitybar": [
-        {
+      "activitybar": [{
           "id": "metadata",
           "title": "Org Browser",
           "icon": "resources/dep.svg"
@@ -102,37 +101,28 @@
       ]
     },
     "views": {
-      "explorer": [
-        {
-          "id": "sfdx.force.tasks.view",
-          "name": "%running_tasks_title_text%"
-        }
-      ],
-      "metadata": [
-        {
-          "id": "metadata",
-          "name": "Metadata",
-          "when": "sfdx:project_opened && sfdx:default_username_has_no_change_tracking"
-        }
-      ],
-      "conflicts": [
-        {
-          "id": "conflicts",
-          "name": "Conflicts",
-          "when": "sfdx:project_opened && sfdx:default_username_has_no_change_tracking"
-        }
-      ]
+      "explorer": [{
+        "id": "sfdx.force.tasks.view",
+        "name": "%running_tasks_title_text%"
+      }],
+      "metadata": [{
+        "id": "metadata",
+        "name": "Metadata",
+        "when": "sfdx:project_opened && sfdx:default_username_has_no_change_tracking"
+      }],
+      "conflicts": [{
+        "id": "conflicts",
+        "name": "Conflicts",
+        "when": "sfdx:project_opened && sfdx:default_username_has_no_change_tracking"
+      }]
     },
     "menus": {
-      "view/title": [
-        {
-          "command": "sfdx.force.metadata.view.type.refresh",
-          "when": "view == metadata",
-          "group": "navigation"
-        }
-      ],
-      "view/item/context": [
-        {
+      "view/title": [{
+        "command": "sfdx.force.metadata.view.type.refresh",
+        "when": "view == metadata",
+        "group": "navigation"
+      }],
+      "view/item/context": [{
           "command": "sfdx.force.task.stop",
           "when": "view == sfdx.force.tasks.view"
         },
@@ -152,8 +142,7 @@
           "group": "inline"
         }
       ],
-      "editor/context": [
-        {
+      "editor/context": [{
           "command": "sfdx.force.source.retrieve.current.source.file",
           "when": "sfdx:project_opened && sfdx:default_username_has_no_change_tracking && resourceLangId != 'forcesourcemanifest'"
         },
@@ -186,8 +175,7 @@
           "when": "sfdx:project_opened && resourceLangId == visualforce"
         }
       ],
-      "explorer/context": [
-        {
+      "explorer/context": [{
           "command": "sfdx.force.lightning.lwc.create",
           "when": "explorerResourceIsFolder && resourceFilename == lwc && sfdx:project_opened"
         },
@@ -280,8 +268,7 @@
           "when": "sfdx:project_opened && sfdx:default_username_has_no_change_tracking && resourceLangId != 'forcesourcemanifest'"
         }
       ],
-      "commandPalette": [
-        {
+      "commandPalette": [{
           "command": "sfdx.force.package.install",
           "when": "sfdx:project_opened"
         },
@@ -511,8 +498,7 @@
         }
       ]
     },
-    "commands": [
-      {
+    "commands": [{
         "command": "sfdx.force.package.install",
         "title": "%force_package_install_text%"
       },
@@ -805,8 +791,7 @@
         }
       }
     },
-    "jsonValidation": [
-      {
+    "jsonValidation": [{
         "fileMatch": "sfdx-project.json",
         "url": "https://forcedotcom.github.io/schemas/sfdx-project.json/sfdx-project.schema.json"
       },
@@ -815,8 +800,7 @@
         "url": "https://forcedotcom.github.io/schemas/project-scratch-def.json/project-scratch-def.schema.json"
       }
     ],
-    "languages": [
-      {
+    "languages": [{
         "id": "ignore",
         "filenames": [
           ".forceignore"
@@ -843,12 +827,10 @@
         ]
       }
     ],
-    "grammars": [
-      {
-        "language": "forcesourcemanifest",
-        "scopeName": "manifest.text.xml",
-        "path": "./syntaxes/manifestXML.tmLanguage.json"
-      }
-    ]
+    "grammars": [{
+      "language": "forcesourcemanifest",
+      "scopeName": "manifest.text.xml",
+      "path": "./syntaxes/manifestXML.tmLanguage.json"
+    }]
   }
 }
